* common.h: Add boost::scoped_array * dictionary.cpp: Add functions words(), tags(), tags_transit() 	Add function add_tag_transit_weight(int tags_index, int weight) * dictionary.h: remove unused types * openclas_vs2005.vcproj: add serialization.{h,cpp} * serialization.h: 	define load_from_file() and save_to_file(); 	Add struct DictHeader, WordHeader, TagItem, TransitHeader 	Add const std::locale locale_utf8 	Add const DICT_MAGIC_CODE * serialization.cpp: implemented load_from_file() and save_to_file() for new dict file format. * utility.cpp: fix a bug(int => size_t)

+++ OpenClas/utility.cpp
@@ -168,7 +168,7 @@ namespace openclas {
 	{
 		const code_converter_type& cc = std::use_facet<code_converter_type>(loc);
 
-		int buf_size = str.length() + 1;
+		size_t buf_size = str.length() + 1;
 		wchar_t* buf = new wchar_t[buf_size];
 
 		mbstate_t state = mbstate_t();
